--- src/commands/insertHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './insertHeader';
